+++ integration-tests/src/test/java/com/transferwise/tasks/ext/management/dao/MySqlManagementTaskDaoIntTest.java
@@ -1,4 +1,5 @@
 package com.transferwise.tasks.ext.management.dao;
 
 public class MySqlManagementTaskDaoIntTest extends ManagementTaskDaoIntTest {
+
 }