+++ integration-tests/src/test/java/com/transferwise/tasks/helpers/TopicPartitionsManagerIntTest.java
@@ -3,7 +3,6 @@ package com.transferwise.tasks.helpers;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.transferwise.tasks.BaseIntTest;
-import com.transferwise.tasks.config.TwTasksKafkaConfiguration;
 import com.transferwise.tasks.helpers.kafka.ITopicPartitionsManager;
 import com.transferwise.tasks.helpers.kafka.NoOpTopicPartitionsManager;
 import java.lang.management.ManagementFactory;
@@ -20,13 +19,11 @@ public class TopicPartitionsManagerIntTest extends BaseIntTest {
 
   @Autowired
   private ITopicPartitionsManager noOpTopicPartitionsManager;
-  @Autowired
-  private TwTasksKafkaConfiguration kafkaConfiguration;
 
   @Test
   public void testAdminClientMbeanId() throws MalformedObjectNameException {
     assertTrue(noOpTopicPartitionsManager instanceof NoOpTopicPartitionsManager);
-    try (AdminClient adminClient = noOpTopicPartitionsManager.createKafkaAdminClient(kafkaConfiguration.getKafkaProperties())) {
+    try (AdminClient ignored = noOpTopicPartitionsManager.createKafkaAdminClient()) {
       MBeanServer mbeanServer = ManagementFactory.getPlatformMBeanServer();
       String objectName = "kafka.admin.client:type=app-info,id=adminclient-*";
       Set<ObjectInstance> objectInstances = mbeanServer.queryMBeans(ObjectNameManager.getInstance(objectName), null);