+++ integration-tests/src/test/java/com/transferwise/tasks/test/dao/MySqlTestTaskDaoIntTest.java
@@ -1,4 +1,5 @@
 package com.transferwise.tasks.test.dao;
 
 public class MySqlTestTaskDaoIntTest extends TestTaskDaoIntTest {
+
 }