+++ tw-tasks-core/src/main/java/com/transferwise/tasks/helpers/kafka/NoOpTopicPartitionsManager.java
@@ -1,9 +1,10 @@
 package com.transferwise.tasks.helpers.kafka;
 
 import com.transferwise.common.baseutils.ExceptionUtils;
-import com.transferwise.tasks.config.TwTasksKafkaConfiguration;
+import com.transferwise.tasks.TasksProperties;
 import java.time.Duration;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
@@ -14,7 +15,6 @@ import org.apache.kafka.clients.admin.DescribeTopicsResult;
 import org.apache.kafka.clients.admin.TopicDescription;
 import org.apache.kafka.common.errors.UnknownTopicOrPartitionException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 
 @Slf4j
 public class NoOpTopicPartitionsManager implements ITopicPartitionsManager {
@@ -22,21 +22,23 @@ public class NoOpTopicPartitionsManager implements ITopicPartitionsManager {
   private static final int COMMANDS_TIMEOUT_S = 30;
 
   @Autowired
-  private TwTasksKafkaConfiguration kafkaConfiguration;
+  private TasksProperties tasksProperties;
 
   @Override
-  public AdminClient createKafkaAdminClient(KafkaProperties kafkaProperties) {
-    Map<String, Object> adminConfig = kafkaConfiguration.getKafkaProperties().buildAdminProperties();
+  public AdminClient createKafkaAdminClient() {
+    Map<String, Object> config = new HashMap<>();
+    config.put(AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG, tasksProperties.getTriggering().getKafka().getBootStrapServers());
     //we are passing empty client-id, this will generate auto increment id with format: "adminclient-" + ADMIN_CLIENT_ID_SEQUENCE.getAndIncrement()
     //see https://github.com/apache/kafka/blob/trunk/clients/src/main/java/org/apache/kafka/clients/admin/KafkaAdminClient.java#generateClientId
-    adminConfig.put(AdminClientConfig.CLIENT_ID_CONFIG, "");
-    return AdminClient.create(adminConfig);
+    config.put(AdminClientConfig.CLIENT_ID_CONFIG, "");
+
+    return AdminClient.create(config);
   }
 
   @Override
   public void setPartitionsCount(String topic, int partitionsCount) {
     ExceptionUtils.doUnchecked(() -> {
-      AdminClient adminClient = createKafkaAdminClient(kafkaConfiguration.getKafkaProperties());
+      AdminClient adminClient = createKafkaAdminClient();
       //noinspection TryFinallyCanBeTryWithResources
       try {
         DescribeTopicsResult describeTopicsResult = adminClient.describeTopics(Collections.singletonList(topic));