+++ tw-tasks-kafka-listener-spring-boot-starter/src/main/java/com/transferwise/tasks/ext/kafkalistener/autoconfigure/TwTasksExtKafkaListenerAutoConfiguration.java
@@ -2,14 +2,17 @@ package com.transferwise.tasks.ext.kafkalistener.autoconfigure;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.transferwise.tasks.ITasksService;
+import com.transferwise.tasks.helpers.kafka.configuration.TwTasksKafkaConfiguration;
 import com.transferwise.tasks.helpers.kafka.messagetotask.CoreKafkaListener;
 import com.transferwise.tasks.helpers.kafka.messagetotask.IKafkaMessageHandlerRegistry;
 import com.transferwise.tasks.helpers.kafka.messagetotask.KafkaMessageHandlerFactory;
 import com.transferwise.tasks.helpers.kafka.messagetotask.KafkaMessageHandlerRegistry;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.core.KafkaTemplate;
 
 @EnableConfigurationProperties
 @Configuration
@@ -33,5 +36,11 @@ public class TwTasksExtKafkaListenerAutoConfiguration {
   @ConditionalOnMissingBean
   public KafkaMessageHandlerFactory kafkaMessageHandlerFactory(ITasksService tasksService, ObjectMapper objectMapper) {
     return new KafkaMessageHandlerFactory(tasksService, objectMapper);
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public TwTasksKafkaConfiguration twTaskKafkaConfiguration(KafkaProperties kafkaProperties, KafkaTemplate<String, String> kafkaTemplate) {
+    return new TwTasksKafkaConfiguration(kafkaProperties, kafkaTemplate);
   }
 }